+++ src/deluge/gui/menu_item/patch_cable_strength.h
@@ -41,9 +41,6 @@ public:
 	ActionResult buttonAction(deluge::hid::Button b, bool on);
 	void selectAutomationViewParameter(bool clipMinder);
 
-	deluge::modulation::params::Kind getParamKind();
-	uint32_t getParamIndex();
-
 	// OLED Only
 	void renderOLED();