+++ src/deluge/gui/ui/sound_editor.cpp
@@ -261,7 +261,6 @@ ActionResult SoundEditor::buttonAction(deluge::hid::Button b, bool on, bool inCa
 				else {
 					goUpOneLevel();
 				}
-				// potentially refresh automation view if entering a new param menu
 				getCurrentMenuItem()->buttonAction(b, on);
 			}
 		}
@@ -282,12 +281,6 @@ ActionResult SoundEditor::buttonAction(deluge::hid::Button b, bool on, bool inCa
 				else {
 					goUpOneLevel();
 				}
-				// if we back out of a patch cable menu
-				// (e.g. LFO -> Velocity -> Level back out to Velocity -> Level)
-				// or if you back out of a patch cable menu back to a regular param menu
-				// (e.g. Velocity -> Level back out to Level)
-				// potentially refresh automation view
-				getCurrentMenuItem()->buttonAction(b, on);
 			}
 		}
 	}
@@ -407,7 +400,6 @@ ActionResult SoundEditor::buttonAction(deluge::hid::Button b, bool on, bool inCa
 	}
 
 	else {
-		// potentially exit out of param / patch cable menu into automation view
 		return getCurrentMenuItem()->buttonAction(b, on);
 	}
 
@@ -467,7 +459,7 @@ bool SoundEditor::findPatchedParam(int32_t paramLookingFor, int32_t* xout, int32
 	for (int32_t x = 0; x < 15; x++) {
 		for (int32_t y = 0; y < kDisplayHeight; y++) {
 			if (paramShortcutsForSounds[x][y] && paramShortcutsForSounds[x][y] != comingSoonMenu
-			    && ((MenuItem*)paramShortcutsForSounds[x][y])->getParamIndex() == paramLookingFor) {
+			    && ((MenuItem*)paramShortcutsForSounds[x][y])->getPatchedParamIndex() == paramLookingFor) {
 
 				*xout = x;
 				*yout = y;
@@ -933,12 +925,6 @@ doSetup:
 					else {
 						display->setNextTransitionDirection(0);
 						beginScreen();
-
-						if (getRootUI() == &automationView) {
-							// if automation view is open in the background
-							// potentially refresh grid if opening a new parameter menu
-							getCurrentMenuItem()->buttonAction(hid::button::SELECT_ENC, on);
-						}
 					}
 				}
 			}