+++ src/deluge/modulation/patch/patch_cable_set.h
@@ -111,8 +111,6 @@ public:
 
 	Destination* destinations[2];
 
-	bool shouldParamIndicateMiddleValue(ModelStackWithParamId const* modelStack) { return true; };
-
 private:
 	static void dissectParamId(uint32_t paramId, ParamDescriptor* destinationParamDescriptor, PatchSource* s);
 	void swapCables(int32_t c1, int32_t c2);